fix(template): fix template bug that can't select by variable

+++ framework/context.go
@@ -20,7 +20,7 @@ const (
 
 // Iterate set iterator variable in context variable
 func Iterate(ctx *runtime.Context, iter int) {
-	ctx.Variables.Set(IteratorName, jsonutil.NewIntVariable(IteratorName, iter))
+	ctx.Variables.Set(IteratorName, jsonutil.NewIntVariable(IteratorName, int64(iter)))
 }
 
 var (
